+++ .env
@@ -1,4 +1,4 @@
 RUST_BACKTRACE=1
 RUST_LOG=debug
 FAUCET_CONFIG=config/faucet.yaml
-TELE_OUT_CONFIG=config/tele_out.yaml
+BLOCK_SUBMITTER_CONFIG=config/block_submitter.yaml
