+++ src/block_submitter/storage.rs
@@ -1,5 +1,5 @@
+use crate::block_submitter::Settings;
 use crate::storage::{PoolOptions, PoolType};
-use crate::tele_out::Settings;
 
 pub async fn from_config(config: &Settings) -> anyhow::Result<PoolType> {
     let db_pool = PoolOptions::new().connect(&config.db).await?;
