+++ src/lib.rs
@@ -1,6 +1,6 @@
+pub mod block_submitter;
 pub mod contracts;
 pub mod faucet;
 pub mod grpc_client;
 pub mod mq;
 pub mod storage;
-pub mod tele_out;
